# valo-mapper-api/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.CancelSubscriptionResponse:
    properties:
      cancelAtPeriodEnd:
        type: boolean
      status:
        type: string
      subscriptionId:
        type: string
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        type: string
      requestId:
        type: string
    type: object
...
  handlers.StripeWebhookResponse:
    properties:
      eventId:
        type: string
      eventType:
        type: string
      reason:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: API for ValoMapper backend services.
  title: ValoMapper API
  version: "1.0"
paths:
...
  /api/billing/cancel-subscription:
    post:
      description: Schedules cancellation at period end for the authenticated user's
        active subscription.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CancelSubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cancel current subscription
      tags:
      - billing
...
  /api/billing/resume-subscription:
    post:
      description: Removes a scheduled cancellation from the authenticated user's
        active subscription.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CancelSubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Resume current subscription
      tags:
      - billing
  /api/billing/webhook:
    post:
      consumes:
      - application/json
      description: Validates and processes Stripe subscription-related webhook events.
      parameters:
      - description: Stripe webhook signature
        in: header
        name: Stripe-Signature
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.StripeWebhookResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Stripe webhook receiver
      tags:
      - billing
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Firebase bearer token in the form "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
  InternalAPIKey:
    description: Internal service API key.
    in: header
    name: X-Internal-API-Key
    type: apiKey
swagger: "2.0"